organizerbotMain/processors/src/organizerbot/processors/watermark.py
"""
Watermark removal functionality for OrganizerBot
"""
from PIL import Image
from organizerbot.core.config import load_config

def remove_watermark(image: Image.Image) -> bool:
    """
    Remove watermark from an image if the feature is enabled
    
    Args:
        image: PIL Image object
        
    Returns:
        bool: True if watermark was removed, False otherwise
    """
    config = load_config()
    if not config.features["watermark_removal"]:
        return False
    
    try:
        # TODO: Implement actual watermark removal logic
        # This is a placeholder for the actual implementation
        return True
    except Exception:
        return False 