requirements.txt
torch>=2.0.0
transformers>=4.30.0
customtkinter>=5.2.0
pillow>=9.5.0
opencv-python>=4.7.0
numpy>=1.24.0
psutil>=5.9.0
watchdog>=3.0.0
pystray>=0.19.0
scikit-learn==1.2.2
joblib==1.2.0
python-telegram-bot>=20.0.0 